Address several kinds of Java compiler warnings (#1361)

+++ optaplanner-benchmark/src/main/java/org/optaplanner/benchmark/impl/report/BenchmarkReport.java
@@ -409,7 +409,7 @@ public class BenchmarkReport {
                         while (solverBenchmarkIndex >= seriesList.size()) {
                             seriesList.add(new XYSeries(solverLabel));
                         }
-                        seriesList.get(solverBenchmarkIndex).add((double) problemScale, levelValues[i]);
+                        seriesList.get(solverBenchmarkIndex).add(problemScale, levelValues[i]);
                     }
                 }
             }
