+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/heuristic/selector/value/decorator/EntityDependentSortingValueSelector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -101,7 +101,7 @@ public class EntityDependentSortingValueSelector<Solution_> extends AbstractValu
     @Override
     public Iterator<Object> iterator(Object entity) {
         long childSize = childValueSelector.getSize(entity);
-        if (childSize > (long) Integer.MAX_VALUE) {
+        if (childSize > Integer.MAX_VALUE) {
             throw new IllegalStateException("The selector (" + this
                     + ") has a childValueSelector (" + childValueSelector
                     + ") with childSize (" + childSize
