+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/ScoreUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -118,7 +118,7 @@ public class ScoreUtils {
                 // timeGradient += 0.0
                 break;
             } else {
-                double levelTimeGradient = (double) scoreDiffLevel / (double) totalDiffLevel;
+                double levelTimeGradient = scoreDiffLevel / totalDiffLevel;
                 timeGradient += levelTimeGradient * levelTimeGradientWeight;
             }
