+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/solver/termination/StepCountTermination.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -65,7 +65,7 @@ public class StepCountTermination<Solution_> extends AbstractTermination<Solutio
     @Override
     public double calculatePhaseTimeGradient(AbstractPhaseScope<Solution_> phaseScope) {
         int nextStepIndex = phaseScope.getNextStepIndex();
-        double timeGradient = ((double) nextStepIndex) / ((double) stepCountLimit);
+        double timeGradient = nextStepIndex / ((double) stepCountLimit);
         return Math.min(timeGradient, 1.0);
     }
