+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/solver/termination/TimeMillisSpentTermination.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -73,7 +73,7 @@ public class TimeMillisSpentTermination<Solution_> extends AbstractTermination<S
     }
 
     protected double calculateTimeGradient(long timeMillisSpent) {
-        double timeGradient = ((double) timeMillisSpent) / ((double) timeMillisSpentLimit);
+        double timeGradient = timeMillisSpent / ((double) timeMillisSpentLimit);
         return Math.min(timeGradient, 1.0);
     }
