+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/solver/termination/UnimprovedStepCountTermination.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -71,7 +71,7 @@ public class UnimprovedStepCountTermination<Solution_> extends AbstractTerminati
     @Override
     public double calculatePhaseTimeGradient(AbstractPhaseScope<Solution_> phaseScope) {
         int unimprovedStepCount = calculateUnimprovedStepCount(phaseScope);
-        double timeGradient = ((double) unimprovedStepCount) / ((double) unimprovedStepCountLimit);
+        double timeGradient = unimprovedStepCount / ((double) unimprovedStepCountLimit);
         return Math.min(timeGradient, 1.0);
     }
